テストのディレクトリ構成を変更しました。
